refactor: ✨ #5 NetworkManagerの画像フェッチ処理を単一責任原則に基づきリファクタリング
- 画像フェッチ処理をGitHubRepositoryImageFetcherクラスに移動
- NetworkManagerから画像フェッチの責任を分離
