+++ iOSEngineerCodeCheck/Services/GitHubRepositoryImageFetcher.swift
@@ -1,0 +1,38 @@
+//
+//  GitHubRepositoryImageFetcher.swift
+//  iOSEngineerCodeCheck
+//
+//  Created by 鈴木斗夢 on 2024/06/17.
+//  Copyright © 2024 YUMEMI Inc. All rights reserved.
+//
+
+import UIKit
+
+class GitHubRepositoryImageFetcher {
+    // シングルトンパターンを利用するためNetworkManagerクラスのインスタン化を禁止
+    private init() {}
+    static let shared = GitHubRepositoryImageFetcher()
+    
+    func fetchRepositoryImage(from urlString: String, completion: @escaping (Result<UIImage?, Error>) -> Void) {
+        guard let imgURL = URL(string: urlString) else {
+            let error = NSError(domain: "Invalid URL", code: -1, userInfo: nil)
+            completion(.failure(error))
+            return
+        }
+        fetchImage(from: imgURL, completion: completion)
+    }
+
+    private func fetchImage(from url: URL, completion: @escaping (Result<UIImage?, Error>) -> Void) {
+        URLSession.shared.dataTask(with: url) { data, _, error in
+            if let error = error {
+                completion(.failure(error))
+                return
+            }
+            if let data = data, let img = UIImage(data: data) {
+                completion(.success(img))
+            } else {
+                completion(.success(nil))
+            }
+        }.resume()
+    }
+}
