+++ iOSEngineerCodeCheck/Services/NetworkManager.swift
@@ -48,27 +48,4 @@ class NetworkManager {
     func cancelSearch() {
         searchTask?.cancel()
     }
-    
-    func fetchRepositoryImage(from urlString: String, completion: @escaping (Result<UIImage?, Error>) -> Void) {
-        guard let imgURL = URL(string: urlString) else {
-            let error = NSError(domain: "Invalid URL", code: -1, userInfo: nil)
-            completion(.failure(error))
-            return
-        }
-        fetchImage(from: imgURL, completion: completion)
-    }
-
-    private func fetchImage(from url: URL, completion: @escaping (Result<UIImage?, Error>) -> Void) {
-        URLSession.shared.dataTask(with: url) { data, _, error in
-            if let error = error {
-                completion(.failure(error))
-                return
-            }
-            if let data = data, let img = UIImage(data: data) {
-                completion(.success(img))
-            } else {
-                completion(.success(nil))
-            }
-        }.resume()
-    }
 }
